Move Gestures options to their own menu under System settings
good lord (cit)
