+++ res/xml/gestures_prefs.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/*
+ * Copyright 2008, The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+-->
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+        android:title="@string/input_methods_settings_title">
+
+    <PreferenceCategory
+        android:key="gesture_settings_category"
+        android:title="@string/gesture_preference_title">
+
+        <Preference
+            android:key="gesture_assist_input_summary"
+            android:title="@string/assist_gesture_title"
+            android:fragment="com.android.settings.gestures.AssistGestureSettings"/>
+
+        <Preference
+            android:key="gesture_swipe_down_fingerprint_input_summary"
+            android:title="@string/fingerprint_swipe_for_notifications_title"
+            android:fragment="com.android.settings.gestures.SwipeToNotificationSettings"/>
+
+        <Preference
+            android:key="gesture_double_tap_power_input_summary"
+            android:title="@string/double_tap_power_for_camera_title"
+            android:fragment="com.android.settings.gestures.DoubleTapPowerSettings"/>
+
+        <Preference
+            android:key="gesture_double_twist_input_summary"
+            android:title="@string/double_twist_for_camera_mode_title"
+            android:fragment="com.android.settings.gestures.DoubleTwistGestureSettings"/>
+
+        <Preference
+            android:key="gesture_double_tap_screen_input_summary"
+            android:title="@string/ambient_display_title"
+            android:fragment="com.android.settings.gestures.DoubleTapScreenSettings"/>
+
+        <Preference
+            android:key="gesture_pick_up_input_summary"
+            android:title="@string/ambient_display_pickup_title"
+            android:fragment="com.android.settings.gestures.PickupGestureSettings"/>
+
+    </PreferenceCategory>
+</PreferenceScreen>
